backend/views/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Mail
Flask-Migrate
Flask-RESTful
Flask-CORS
Pillow
mysqlclient
python-dotenv
werkzeug